refactor: remove _transition_info module
