+++ src/ampform/helicity.py
@@ -8,23 +8,16 @@ from typing import Any, Dict, Iterable, List, Optional, Tuple, Type, Union
 import attr
 import sympy as sp
 from attr.validators import instance_of
-from qrules import ParticleCollection, ReactionInfo
 from qrules.combinatorics import (
     perform_external_edge_identical_particle_combinatorics,
 )
+from qrules.particle import ParticleCollection, Spin
 from qrules.quantum_numbers import InteractionProperties
-from qrules.transition import State, StateTransition
+from qrules.transition import ReactionInfo, State, StateTransition
 from sympy.physics.quantum.cg import CG
 from sympy.physics.quantum.spin import Rotation as Wigner
 from sympy.printing.latex import LatexPrinter
 
-from ._transition_info import (
-    generate_particle_collection,
-    get_angular_momentum,
-    get_coupled_spin,
-    get_prefactor,
-    group_transitions,
-)
 from .dynamics.builder import (
     ResonanceDynamicsBuilder,
     TwoBodyKinematicVariableSet,
@@ -479,7 +472,7 @@ class HelicityAmplitudeBuilder:  # pylint: disable=too-many-instance-attributes
         self.__adapter = HelicityAdapter(reaction)
         for grouping in reaction.transition_groups:
             self.__adapter.register_topology(grouping.topology)
-        self.__particles = generate_particle_collection(reaction.transitions)
+        self.__particles = extract_particle_collection(reaction.transitions)
 
     def set_dynamics(
         self, particle_name: str, dynamics_builder: ResonanceDynamicsBuilder
@@ -731,9 +724,111 @@ def _latex_fix(self: Type[CG], printer: LatexPrinter, *args: Any) -> str:
 CG._latex = _latex_fix
 
 
+def extract_particle_collection(
+    transitions: Iterable[StateTransition],
+) -> ParticleCollection:
+    particles = ParticleCollection()
+    for transition in transitions:
+        for state in transition.states.values():
+            if state.particle not in particles:
+                particles.add(state.particle)
+    return particles
+
+
+def get_angular_momentum(interaction: InteractionProperties) -> Spin:
+    l_magnitude = interaction.l_magnitude
+    l_projection = interaction.l_projection
+    if l_magnitude is None or l_projection is None:
+        raise TypeError(
+            "Angular momentum L not defined!", l_magnitude, l_projection
+        )
+    return Spin(l_magnitude, l_projection)
+
+
+def get_coupled_spin(interaction: InteractionProperties) -> Spin:
+    s_magnitude = interaction.s_magnitude
+    s_projection = interaction.s_projection
+    if s_magnitude is None or s_projection is None:
+        raise TypeError("Coupled spin S not defined!")
+    return Spin(s_magnitude, s_projection)
+
+
+def get_prefactor(
+    transition: StateTransition,
+) -> float:
+    """Calculate the product of all prefactors defined in this transition."""
+    prefactor = 1.0
+    for node_id in transition.topology.nodes:
+        interaction = transition.interactions[node_id]
+        if interaction:
+            temp_prefactor = __validate_float_type(
+                interaction.parity_prefactor
+            )
+            if temp_prefactor is not None:
+                prefactor *= temp_prefactor
+    return prefactor
+
+
+def group_transitions(
+    transitions: Iterable[StateTransition],
+) -> List[List[StateTransition]]:
+    """Match final and initial states in groups.
+
+    Each `~qrules.transition.StateTransition` corresponds to a specific state
+    transition amplitude. This function groups together transitions, which have the
+    same initial and final state (including spin). This is needed to determine
+    the coherency of the individual amplitude parts.
+    """
+    transition_groups: Dict[Tuple[tuple, tuple], List[StateTransition]] = {}
+    for transition in transitions:
+        initial_state_ids = transition.topology.outgoing_edge_ids
+        final_state_ids = transition.topology.incoming_edge_ids
+        transition_group = (
+            tuple(
+                sorted(
+                    [
+                        (
+                            transition.states[i].particle.name,
+                            transition.states[i].spin_projection,
+                        )
+                        for i in initial_state_ids
+                    ]
+                )
+            ),
+            tuple(
+                sorted(
+                    [
+                        (
+                            transition.states[i].particle.name,
+                            transition.states[i].spin_projection,
+                        )
+                        for i in final_state_ids
+                    ]
+                )
+            ),
+        )
+        if transition_group not in transition_groups:
+            transition_groups[transition_group] = []
+        transition_groups[transition_group].append(transition)
+
+    return list(transition_groups.values())
+
+
 def __attempt_int_cast(value: float) -> Union[float, int]:
     if isinstance(value, int):
         return value
     if value.is_integer():
         return int(value)
     return value
+
+
+def __validate_float_type(
+    interaction_property: Optional[Union[Spin, float]]
+) -> Optional[float]:
+    if interaction_property is not None and not isinstance(
+        interaction_property, (float, int)
+    ):
+        raise TypeError(
+            f"{interaction_property.__class__.__name__} is not of type {float.__name__}"
+        )
+    return interaction_property
